+++ wpcleanadmin/includes/class-wpca-database-settings.php
@@ -1,4 +1,4 @@
-﻿<?php
+<?php
 /**
  * WP Clean Admin Database Settings
  * 
@@ -7,7 +7,7 @@
  * 
  * @package WP_Clean_Admin
  * @since 1.6.0
- * @version 1.7.12
+ * @version 1.7.13
  */
 
 if ( ! defined( 'ABSPATH' ) ) {